#8 int to Long

+++ src/main/kotlin/com/example/kotlinPractice/feature/menu/api/dto/MenuCreateClientDto.kt
@@ -2,5 +2,5 @@ package com.example.kotlinPractice.feature.menu.api.dto
 
 data class MenuCreateClientDto(
     val menuCreateDto: MenuCreateDto,
-    val kitchenId: Int,
+    val kitchenId: Long,
 )
